Don't throw NPE on GoogleCredential.Builder when scopes not set
The list of account scopes in the GoogleCredential.Builder starts off
uninitialized (i.e. null), meaning that calling the .build() method
throws an NPE unless setServiceAccountScopes() is called.  There are no
other fields in the builder that are required to be called, and this
doesn't seem to be required intentionally, so make the Builder resilient
to it being null and have it correctly interpret that as an empty list
of scopes.