Add missing "fall through" comments to switches that fall through
The Style Guide requires that within a switch block, each statement group
either terminates abruptly (e.g. with a break), or is marked with a comment to
indicate that execution may "fall through" to next statement group.

This change is being made in preparation for enabling a new compiler error,
which will ensure a "fall through" comment is present if and only if the switch
falls through.

NOTE TO REVIEWERS:

Please review carefully to ensure the fall through is deliberate. If it is a
bug and there should e.g. be a 'break;' statement, please respond with a
suggested edit.

More information:
https://docs.google.com/a/google.com/document/d/1cz7o7p9ROfXZ0VMtUfFN7Tnw3F61qkFw9bnJ64GDiP8/edit?usp=sharing

Tested:
    TAP --sample for global presubmit queue
    http://test/OCL:147852686:BASE:147857215:1487376286182:2ba055dd

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=148101033