add jdo jar properties to android assembly (#1131)
When the assembly project is {compiled|deployed} (one of those two?), it generates sources. Among those sources are the google-http-java-client-jdo-<version>.jar file in libs-sources. This is fine, but the release process expects a .jar.properties file in libs/. This commit adds it.

Prior to this commit, people were scratching their heads during release and adding it manually.

ejona@ was the one that discovered the solution to this problem.