docs: Update README.md - update CI status badge
The previous badges were pointing to old CI (last run in 2021) leading
to false impression that it is tested with Java 7, 8 and 11 only.

I updated to point to new GitHub Ci. I only pointed to the top-level as
I have not found a way to point to specific job of a GitHub workflow.